Fixed issue with GraphQL
