+++ Starwars/GraphQL/GraphQL.graphql.swift
@@ -1,0 +1,4 @@
+// @generated
+// This file was automatically generated and should not be edited.
+
+public enum GraphQL { }
