+++ Starwars/GraphQL/Schema/Objects/Planet.graphql.swift
@@ -3,11 +3,11 @@
 
 import ApolloAPI
 
-public extension Objects {
+public extension GraphQL.Objects {
   /// A large mass, planet or planetoid in the Star Wars Universe, at the time of
   /// 0 ABY.
   static let Planet = Object(
     typename: "Planet",
-    implementedInterfaces: [Interfaces.Node.self]
+    implementedInterfaces: [GraphQL.Interfaces.Node.self]
   )
 }
