+++ Starwars/GraphQL/Schema/Objects/Vehicle.graphql.swift
@@ -3,10 +3,10 @@
 
 import ApolloAPI
 
-public extension Objects {
+public extension GraphQL.Objects {
   /// A single transport craft that does not have hyperdrive capability
   static let Vehicle = Object(
     typename: "Vehicle",
-    implementedInterfaces: [Interfaces.Node.self]
+    implementedInterfaces: [GraphQL.Interfaces.Node.self]
   )
 }
