+++ Starwars/GraphQL/Schema/SchemaConfiguration.swift
@@ -7,8 +7,8 @@
 
 import ApolloAPI
 
-public enum SchemaConfiguration: ApolloAPI.SchemaConfiguration {
-  public static func cacheKeyInfo(for type: Object, object: JSONObject) -> CacheKeyInfo? {
+enum SchemaConfiguration: ApolloAPI.SchemaConfiguration {
+  static func cacheKeyInfo(for type: Object, object: JSONObject) -> CacheKeyInfo? {
     // Implement this function to configure cache key resolution for your schema types.
     return nil
   }
